requirements.txt
arabic-reshaper==2.1.3
BD==0.0.5
certifi==2021.5.30
charset-normalizer==2.0.5
cloudinary==1.26.0
dj-database-url==0.5.0
Django==2.0.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
future==0.18.2
gunicorn==19.9.0
html5lib==1.1
idna==3.2
Pillow==8.3.2
psycopg2==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.6.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
